fix(webhook): always re-apply modifications in instrumentation webhook
Previously, the webhook would check whether a workload had already been
instrumented by the same operator version, and, if so, would not modify
the workload again. The same logic is used when instrumenting existing
workloads.

However, in the webhook, this conditional application of modifications
is actually detrimental. In particular, if the workload is modified in a
way that makes it invalid in combination with the Dash0 workload
modifications, the update is rejected.

Example: The user deletes all volumes from the workload. This will
result in an invalid state because the pods still have the
dash0-instrumentation volume mounts.

Instead of ignoring that request in the instrumentation_webhook, which
then leads to Kubernetes ultimately rejecting the workload update, the
better strategy is to always re-apply our modifications.

Furthermore, the only reason for not applying the modifications again is
to not restart pods, when modifying _existing_ workloads. But this
aspect is irrelevant in the webhook (which only is triggered when
workloads are modified anyway).

We already made this change a month ago, in
7890dcdd3c476dddd0d1400fd5862462bf331386, but needed to revert it with
e4665b68c180c64b513651ee3b81ccdcbfc33b11 due to a conflict with how we
handled OTEL_EXPORTER_OTLP_ENDPOINT/OTEL_EXPORTER_OTLP_PROTOCOL
modifications. In particular, workload_modifier.go would always
overwrite the _ENDPOINT with our collector's HTTP endpoint (and set the
_PROTOCOL accordingly), even if the container already defines the
_ENDPOINT. This is problematic for containers that implicitly use a gRPC
export and ignore OTEL_EXPORTER_OTLP_PROTOCOL.

This issue was hidden by our previous webhook modification logic, i.e.
an IaC system would re-apply its idea of OTEL_EXPORTER_OTLP_ENDPOINT on
the workload after the operator's modifications, and telemetry would
flow correctly. This self-healing broke with
7890dcdd3c476dddd0d1400fd5862462bf331386.

Now, with the changes from
https://github.com/dash0hq/dash0-operator/pull/690, we can restore the
webhook behavior change from 7890dcdd3c476dddd0d1400fd5862462bf331386,
which is correct on its own and also needed for the scenarios described
above.

This reverts commit e4665b68c180c64b513651ee3b81ccdcbfc33b11.